html/semantics/document-metadata/the-base-element/base_href_data.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/get-host-info.sub.js"></script>

  <base href="data:/,This is a data URL.">
</head>
<body>
  <script>
    async_test(t => {
      var base = document.querySelector('base');
      var img = new Image();
      img.onload = t.step_func_done(_ => {
        assert_equals(img.naturalWidth, 256, "Image loaded correctly.");
        assert_equals(img.src, get_host_info().HTTP_ORIGIN + "/images/green-256x256.png");
        assert_equals(base.href, 'data:/,This is a data URL.');
      });
      img.onerror = t.unreached_func("Image should have loaded.");

      img.src = "/images/green-256x256.png";
    }, "'data:' is an invalid base URL.");
  </script>
</body>
